+++ tests/stubs/helix/commands.scm
@@ -2,11 +2,13 @@
 ;; `make compile-check`. See ../README.md.
 
 (provide debug-start
+         new
          write
          open
          goto-line)
 
 (define (debug-start template . arguments) void)
+(define (new . arguments) void)
 (define (write . arguments) void)
 (define (open . arguments) void)
 (define (goto-line line) void)